Фикс бд

+++ database/migrations/2021_02_04_203117_create_currencies_table.php
@@ -16,7 +16,7 @@ class CreateCurrenciesTable extends Migration
     Schema::create('currencies', function (Blueprint $table) {
       $table->id();
       $table->string('name');
-      $table->float('ratio');
+      $table->decimal('ratio', 10, 5);
       $table->string('symbol');
       $table->string('short_name');
       $table->timestamps();
